+++ pkg/local_object_storage/metabase/lock_test.go
@@ -169,6 +169,27 @@ func TestDB_Lock(t *testing.T) {
 	})
 }
 
+func TestDB_Lock_Expired(t *testing.T) {
+	es := &epochState{e: 123}
+
+	db := newDB(t, meta.WithEpochState(es))
+
+	// put an object
+	addr := putWithExpiration(t, db, object.TypeRegular, 124)
+
+	// expire the obj
+	es.e = 125
+	_, err := metaGet(db, addr, false)
+	require.ErrorIs(t, err, meta.ErrObjectIsExpired)
+
+	// lock the obj
+	require.NoError(t, db.Lock(addr.Container(), oidtest.ID(), []oid.ID{addr.Object()}))
+
+	// object is expired but locked, thus, must be available
+	_, err = metaGet(db, addr, false)
+	require.NoError(t, err)
+}
+
 func TestDB_IsLocked(t *testing.T) {
 	db := newDB(t)